Renamed portsentry_io.[ch] and portsentry_util.[ch] to io/util
